+++ src/v/security/authorizer.cc
@@ -11,6 +11,7 @@
 
 #include "security/authorizer.h"
 
+#include "acl_store.h"
 #include "security/role.h"
 #include "security/role_store.h"
 
@@ -19,6 +20,56 @@
 
 namespace security {
 
+authorizer::~authorizer() = default;
+
+authorizer::authorizer(
+  config::binding<std::vector<ss::sstring>> superusers, const role_store* roles)
+  : authorizer(allow_empty_matches::no, std::move(superusers), roles) {}
+
+authorizer::authorizer(
+  allow_empty_matches allow,
+  config::binding<std::vector<ss::sstring>> superusers,
+  const role_store* roles)
+  : _store(std::make_unique<acl_store>())
+  , _superusers_conf(std::move(superusers))
+  , _allow_empty_matches(allow)
+  , _role_store(roles) {
+    update_superusers();
+    _superusers_conf.watch([this]() { update_superusers(); });
+}
+
+void authorizer::add_bindings(const std::vector<acl_binding>& bindings) {
+    if (unlikely(
+          seclog.is_shard_zero() && seclog.is_enabled(ss::log_level::debug))) {
+        for (const auto& binding : bindings) {
+            vlog(seclog.debug, "Adding ACL binding: {}", binding);
+        }
+    }
+    store().add_bindings(bindings);
+}
+
+std::vector<std::vector<acl_binding>> authorizer::remove_bindings(
+  const std::vector<acl_binding_filter>& filters, bool dry_run) {
+    return store().remove_bindings(filters, dry_run);
+}
+
+std::vector<acl_binding>
+authorizer::acls(const acl_binding_filter& filter) const {
+    return store().acls(filter);
+}
+
+ss::future<fragmented_vector<acl_binding>> authorizer::all_bindings() const {
+    return store().all_bindings();
+}
+
+ss::future<>
+authorizer::reset_bindings(const fragmented_vector<acl_binding>& bindings) {
+    return store().reset_bindings(bindings);
+}
+
+acl_store& authorizer::store() & { return *_store; }
+const acl_store& authorizer::store() const& { return *_store; }
+
 template<typename T>
 auth_result authorizer::authorized(
   const T& resource_name,
@@ -26,7 +77,7 @@ auth_result authorizer::authorized(
   const acl_principal& principal,
   const acl_host& host) const {
     auto type = get_resource_type<T>();
-    auto acls = _store.find(type, resource_name());
+    auto acls = store().find(type, resource_name());
 
     if (_superusers.contains(principal)) {
         return auth_result::superuser_authorized(
@@ -54,7 +105,7 @@ auth_result authorizer::authorized(
           "present");
         const acl_principal_base& to_check = *role.value_or(&user);
         bool is_allow = perm == acl_permission::allow;
-        std::optional<acl_matches::acl_match> entry;
+        std::optional<security::acl_match> entry;
         if (is_allow) {
             entry = acl_any_implied_ops_allowed(
               acls, to_check, host, operation);
@@ -164,5 +215,47 @@ template auth_result authorizer::authorized(
   acl_operation,
   const acl_principal&,
   const acl_host&) const;
+
+std::optional<security::acl_match> authorizer::acl_any_implied_ops_allowed(
+  const acl_matches& acls,
+  const acl_principal_base& principal,
+  const acl_host& host,
+  const acl_operation operation) const {
+    auto check_op = [&acls, &principal, &host](
+                      auto begin,
+                      auto end) -> std::optional<security::acl_match> {
+        for (; begin != end; ++begin) {
+            if (auto entry = acls.find(
+                  *begin, principal, host, acl_permission::allow);
+                entry.has_value()) {
+                return entry;
+            }
+        }
+
+        return {};
+    };
+
+    switch (operation) {
+    case acl_operation::describe: {
+        static constexpr std::array ops = {
+          acl_operation::describe,
+          acl_operation::read,
+          acl_operation::write,
+          acl_operation::remove,
+          acl_operation::alter,
+        };
+        return check_op(ops.begin(), ops.end());
+    }
+    case acl_operation::describe_configs: {
+        static constexpr std::array ops = {
+          acl_operation::describe_configs,
+          acl_operation::alter_configs,
+        };
+        return check_op(ops.begin(), ops.end());
+    }
+    default:
+        return acls.find(operation, principal, host, acl_permission::allow);
+    }
+}
 
 } // namespace security
